requirements.txt
-i https://pypi.org/simple
argcomplete==1.11.0
attrs==19.3.0
beancount==2.2.3
beautifulsoup4==4.8.2
bottle==0.12.18
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth==1.10.0
httplib2==0.15.0
idna==2.8
lxml==4.4.2
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
ply==3.11
py==1.8.1
pyasn1-modules==0.2.7
pyasn1==0.4.8
pycparser==2.19
pyparsing==2.4.6
pytest==5.3.2
python-dateutil==2.8.1
python-magic==0.4.15
python-telegram-bot==12.2.0
requests==2.22.0
rsa==4.0
six==1.13.0
soupsieve==1.9.5
tornado==6.0.3
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.8
webdavclient3==0.14
